requirements.txt
beautifulsoup4==4.13.3
dashscope==1.20.14
langchain==0.3.20
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_core==0.3.41
langchain_openai==0.3.7
Markdown==3.4.4
PyQt5==5.15.11
PyQt5_sip==12.16.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
tiktoken==0.9.0
transformers==4.48.2
